Module webservice::models

Structs

InsertablePhoto
Insertable Photo struct (representation of the photos table without the id that is inserted automatically)

InsertableSite
Insertable Site struct (representation of the sites table without the id that is inserted automatically)

InsertableSubmission
Insertable Submission struct (representation of the submissions table without the id that is inserted automatically)

InsertableUser
Insertable User struct (representation of the users table without the id that is inserted automatically)

Photo
Photo struct (representation of the photos table)

Site
Site struct (representation of the sites table)

Submission
Submission struct (representation of the submissions table)

User
User struct (representation of the users table)

Functions

_impl_associations_for_insertablephoto

_impl_associations_for_insertablesubmission

_impl_associations_for_photo

_impl_associations_for_submission

_impl_identifiable_for_photo

_impl_identifiable_for_site

_impl_identifiable_for_submission

_impl_identifiable_for_user

_impl_insertable_for_insertablephoto

_impl_insertable_for_insertablesite

_impl_insertable_for_insertablesubmission

_impl_insertable_for_insertableuser

_impl_queryable_for_photo

_impl_queryable_for_site

_impl_queryable_for_submission

_impl_queryable_for_user